1. action_model_trainer.py: fixed unsupported cpu op 2. max_ent_dqn.py: minor dtype change 3. rooms.py: working on enabling a continuous action space

+++ action_model_trainer.py
@@ -36,4 +36,4 @@ class ActionModelTrainer:
         logger.record("action model/loss", loss.item())
         logger.record("action model/accuracy", acc)
         logger.record("action model/entropy", torch.mean(m.entropy()).item())
-        logger.record("action model/hist", torch.histc(batch.actions, bins=action_probs.shape[1]).tolist())
+        logger.record("action model/hist", torch.histc(batch.actions.float(), bins=action_probs.shape[1]).tolist())
